external links are loaded as soon as possible, important for very first start
